Add Logic for < Level 99 and one button meds.
Adds Logic to swap in to all level gear when under level 99, useful
for event BCNM and level sync leveling. Will swap into a TP50 set
and WS in that set.

Adds a command and one button (F10) med usage if/elseif statement
to use the appropriate med in decending order:

1. Paralyze (Panacea)
2. Doom / Curse (Hallowed Water)
3. Silence (Echo Drops)
4. All other effects that Panacea can remove (Panacea)